GET /remote_assets/ returns only objects 'in progress' (queued or processing)
we still need to decide what we want the API to return  - if we want it to return an array of ids only then that will be tricky with activeresource.
if we want it to return an abbreviated array of hashes like {:remote_asset => {:id => id}, etc} then that's easier with activeresource but i don't know how friendly that is.